Prototype of pre-processing to crop book regions.
Some simple pre processing to cropout ROI before input to a segment model.
In final deployment, this will be displayed as preview in realtime to the users, help them to adjust the camera appropriately.
Accept one row only, expect user to take a photo of book region in one bookshelf row at a time, which covers most of the camera view.
The preview shown to user is also the crop input to the segment model, help model to focus more on the book region.